Users method functionality and compatability update
Commented out the body of the addCourses method and updated it with a
new body that iterates through the arrays passed as arguments.
Updated the set(Phone/Email/Name) methods so the arguments have
different names from the variables being used, and added the framework
for handling invalid arguments when calling this method.